pyproject.toml
[project]
name = "picture-classifier"
version = "0.1.0"
description = "Score photos for blur/exposure/face-clustering and cull via local web viewer"
readme = "README.md"
license = { file = "LICENSE" }
authors = [
    { name = "Hyoungseo Son", email = "son.engr.kr@gmail.com" }
]
keywords = ["photography", "culling", "blur-detection", "face-clustering", "triage"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Environment :: Web Environment",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics",
]
requires-python = ">=3.12"
dependencies = [
    "click>=8.1",
    "fastapi>=0.115",
    "insightface>=0.7.3",
    "numpy>=2.0",
    "onnxruntime>=1.25.0",
    "opencv-python>=4.10",
    "pillow>=10.4",
    "scikit-learn>=1.8.0",
    "uvicorn[standard]>=0.30",
]

[project.optional-dependencies]
faces = ["mediapipe>=0.10"]

[project.scripts]
pcls = "picture_classifier.cli:main"

[project.urls]
Homepage = "https://github.com/son-engr-kr/picture-classifier"
Issues = "https://github.com/son-engr-kr/picture-classifier/issues"

[build-system]
requires = ["uv_build>=0.11.7,<0.12.0"]
build-backend = "uv_build"
